requirements.txt
flask==2.3.3
tensorflow==2.15.0
numpy==1.24.4
Pillow==10.2.0
gunicorn==21.2.0
